try to handle void * to enum cast on Windows

+++ internal/gumbo/error.c
@@ -79,8 +79,8 @@ static void print_tag_stack(const GumboParserError* error, GumboStringBuffer* ou
     if (i) {
       print_message(output, ", ");
     }
-    // cast to long first to prevent void ptr to enum cast warning
-    GumboTag tag = (GumboTag)( (long) (error->tag_stack.data[i]) );
+    // cast to uintptr_t first to prevent void ptr to enum cast warning
+    GumboTag tag = (GumboTag)( (uintptr_t) (error->tag_stack.data[i]) );
     print_message(output, gumbo_normalized_tagname(tag));
   }
   gumbo_string_buffer_append_codepoint('.', output);
